Remove complicated previousAppId usage
Currently, previousAppId being < 0, == 0, or > 0 has different meanings,
which is pretty confusing. We will be keeping track of an appId's
lineage, so the caller should always supply the actual previous appId if
an appId transition happens.

This commit removes this existing complication, and will only perform
appId migration if and only if a valid previousAppId argument is passed.

Test: m installd
Bug: 220015249
Change-Id: I0c52dad86c0e8d7442e659b2fad7a95ac610f774