libbinder: Use file operations in RpcTransportRaw
libbinder RPC assumes its file descriptors are sockets and
uses sendmsg/recvmsg to transfer data in RpcTransportRaw.
This switches the transport to readv/writev so it can be used
on other kinds of file descriptors.

sendmsg accepts the MSG_NOSIGNAL to disable SIGPIPE but writev
does not support this. Instead, we disable the signal manually
for every RPC thread using pthread_sigmask.

Bug: 224064243
Test: atest binderRpcTest
Change-Id: I92057f108fa1478f8ac2085b84e49a36edb00233