Bind mount misc storage to mirror for different volumes
A mirror for misc_ce and misc_de storage is created for each new volume
UUID and a bind mount occurs from /mnt/expand/<uuid>/misc_{ce,de} to
/data_mirror/misc_{ce,de}/<uuid>.

This is similar to how apps use data_mirror. In addition, when a private
volume is unmounted, the mirror for that volume UUID is deleted.

Bug: 214241165
Test: atest SdkSandboxStorageHostTest

Change-Id: I3ea99145e38ed3781a701c1305720f2d4dbe54ce