Anr no focus window for the focused application
For multi TaskDisplayAreas devices when an app is taking too much to
load the user does not receive an Anr dialog because the previous
focused window is still visible and focusable.

Solution: Check if the focused app has the same app token as the
focused window and the focused window is not BASE_APPLICATION type
The possibility of multi window allows us to have the case where
the focused app has no window and FocusResolver picks the next
available target for the window focus and in result no anr is shown
for unresponsive focused
app.

Bug: 231764112
Test: cts-tradefed run commandAndExit cts
-m CtsWindowManagerDeviceTestCases
--test android.server.wm.AnrTests#slowOnCreateWithKeyEventTriggersAnr
Verified on: Pixel 4 & development device

Change-Id: I340a974b84aeea4ca2e84755329441c1a924ca6d