binder: Use variant for backend specific Parcel fields
This frees up some space to add RPC binder specific bookkeeping to
Parcel. `variant` has a size overhead of one pointer, but since there is
at least one pointer worth of fields in each case, it works out as an
overall win. We could probably do it cheaper by using a union and hiding
a bit field somewhere to indicate the active case, but that isn't
necessary yet.

Test: TH
Change-Id: Id53ea0cbfe51b3246e6f9f1f3d9df7c8c193808e