libbinder: use the same RpcTransport for new sessions in RpcServer
RpcServer previously created new RpcSession instances
using the default transport. This CL changes RpcServer to use
the same RpcTransportCtxFactory it was created with, so
the new sessions all use the same transport as the server.

This requires that the RpcTransportCtxFactory is saved inside
the RpcServer, so factory instances are now wrapped using
std::shared_ptr instead of unique pointers.

Bug: 224644083
Change-Id: Ie9fa8c9c758821614fb161b360a9935955e7e62b