libbinder: run testSupportVsockLoopback in a separate process
Split the testSupportVsockLoopback helper function from
binderRpcTest into a separate process so that the client and
service do not require multiple threads in the current process.
Rewrites the service to use socket operations directly instead
of the Rpc* classes so that the out-of-process service can
be cleanly shut down over a pipe.

Bug: 224644083
Test: atest binderRpcTest
Change-Id: Ic30b5d91f72fb961b88ec4f9ba467daaf8c0a24a