libbinder: ignore POLLERR when polling after connect
RpcSession calls triggerablePoll after each non-blocking connect
to check if the connection succeeded. If it did not, poll returns
both POLLOUT and POLLERR. This adds a flag to triggerablePoll to
ignore that POLLERR when called from RpcSession, since the latter
calls getsockopt with SO_ERROR anyway to get the actual error code.

Bug: 224644083
Test: atest binderRpcTest
Change-Id: I16599727ff7355b028a319170a5fb769eddf667e