binderLibTest: wait longer to start all threads
1ms didn't give enough time for all the threads to run. 100ms should
be enough, but honestly it introduces a potential flake. We could try
to merge this and see the results. 100ms or 500ms should be enough
time. Usually tests will wait on a condition variable with a timeout
of this time, though in this case the corresponding condition variable
is not exposed.

Fixes: 233226955
Test: binderLibTest --gtest_filter="*ThreadPoolAvailableThreads*" --gtest_repeat=1000 --gtest_break_on_failure
Change-Id: I937f2d3560d0f1fe01c0c6e0ed140d2d6a823492