Add a delegate helper for AIDL Delegators
This will check if the binder already has/is a delegator by looking in the
attached objects and attempting to promote the object with the right
objectID.
If the object is already a delegator, it will be unwrapped.
If the object already has a delegator, the delegator will be returned
instead.
If neither of those is true, it will create new delegator that wraps the
binder and return the new delegator.

Test: atest binderUnitTest aidl_integration_test
Bug: 220141324
Change-Id: I2c8baa211300fa3ada50f2b2c93862f8c968be26