Specify AIDL filegroups directly in libgui
Currently, `framework_native_aidl` wraps `framework_native_aidl_binder` and `framework_native_aidl_gui`. When converting cc modules with AIDL srcs to Bazel, we run into some challenges with this pattern because an AIDL filegroup module in Soong doesn't map 1-to-1 an aidl_library target in Bazel. See go/bp2build-aidl-filegroup for more details.

Instead of pursuing a more sophisticated solution to handle nested AIDL filegroups, we're proposing to flatten the filegroups directly in Android.bp since scale of the problem is small.

Doing this makes it easier to convert AIDL filegroups to Bazel for cc rules to use and doesn't change anything in how Soong builds the module.

Let me know if there is any drawback of this approach from your perspective.

Test: m libgui
Change-Id: If3c57acbcace2256d369f31a02f78601bfc9eef2