Specify AIDL filegroups directly in libgui
When converting cc modules with AIDL srcs to Bazel, we run into some challenges when a filegroup with AIDL files is nested in another filegroup. See go/bp2build-aidl-filegroup for more details.

Instead of pursuing a sophisticated solution to handle nested AIDL filegroups, we're proposing to flatten the filegroups directly in Android.bp since scale of the problem is small. Doing this make our design for Bazel rules easier to implement and doesn't change anything in how Soong builds the module.

Let me know if there is any drawback of this approach from your perspective.

Test: m libgui
Change-Id: If3c57acbcace2256d369f31a02f78601bfc9eef2