Revert "Factor out helper method."
This is an Android Culprit Assistant revert, created as part of a culprit search session (go/aca-get/81225755-139a-41b5-ad65-308fb3b673bf). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ic0ee38a18de162f74a40fab985e59d16eb79e888