binderRpcTest: wait indefinitely for death recipient
This has been seen timing out occasionally in presubmits, specifically
instances like: SingleDeathRecipient/inet_socket_tls_clientV0_serverV0_no_kernel

I was able to reproduce it on host with a full build running on host. I
could not reproduce it with an idle host or on device locally.

Rather than guessing a time for the wait time, just wait indefinitely
and let the testing infra catch a deadlock if it occurs.

Test: for i in `seq 1 100` ; do ./binderRpcTestNoKernel --gtest_filter=*Death*no_kernel*; done
Bug: NA
Change-Id: I42ad6eb66d36759193980e1532ac9e6ec7565435