Revert "Remove unused variables."
This is an Android Culprit Assistant revert, created as part of a culprit search session (go/aca-get/127bccd7-e848-4b8f-bde4-c4f0ae4b2938). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I790d365be0db68d94fbd507e39cfb7bd3db92a85