Revert "Remove unused variables."
This is an Android Culprit Assistant revert, created as part of a culprit search session (go/aca-get/7124b570-bba8-4e08-9019-06816306c3eb). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ia9242757b3238058e105be21cc66f0527043a50d