Replace libmeminfo tools with bugreport_procdump
This change replaces calls to showmap, procrank, and librank in
dumpstate with a single call to bugreport_procdump, which collects the
same information with less overhead.

Over 3 trials on aosp_redfin-userdebug, the time spent collecting memory
info during a bug report was reduced by ~88% (11.954s -> 1.393s).

With a high outlier removed from the 'before' trials and a low outlier
removed from the 'after' trials, the reduction in overall bug report
time was measured at ~13.5% (50.639s -> 43.784s).

Test: Checked that bug reports still contain the same information, and
      that showmap/procrank/librank sections are still parsed by ABT in
      the same way.
Bug: 229147699
Change-Id: I724922d2c583f2d1b3a443f86d445e994260c1cd