Revert "atrace: add ufshcd_command event and change block events..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (go/aca-get/24b0e4e1-ee7d-492e-812a-ea9c5f5adc25). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Id8e1f41191d610b724c31c241a25929c57cff9d2