BufferQueueConsumer: acquire a buffer only if fence is signaled
BufferQueueConsumer::acquireBuffer() returns the next buffer from the
queue with the associated fence, expecting the caller (typically
GLConsumer.updateTexImage) to wait on the fence before reading from the
buffer (as it may not be ready yet).  This isn't ideal for a
latency-sensitive VR compositor thread -- we would rather keep using the
previous buffer than delay subsequent draw calls.  This diff adds a flag
that causes acquireBuffer() to return NO_BUFFER_AVAILABLE if the fence
hasn't signaled (in which case updateTexImage() keeps the previous
buffer). The flag is exposed via a method on IGraphicBufferProducer,
which apps invoke indirectly via a native window perform() hook.

Test: manual - on Meta Quest 2, add code to VR compositor to simulate
a GPU-intensive application rendering to an Android Surface swap chain.
Observe heavy judder and tearing without this change, resulting from
long delays between the compositor's command batches being queued and
submitted.  Verify issue is resolved with this diff.  On bluejay, verify
device boots and apps are visible and interactable, and no new failures
are introduced in the surfaceflinger test suite.

Change-Id: Ia6a3f3b26e43e076f4acad099f6785b3c6c4a045