Remove bufferhub support from libgui.
As far as I can tell the bufferhub support in libgui was never completed,
and it is unused except for tests and the libdvr library (which is also
unused except for benchmarks and tests, which I also removed). This will
allow building the platform with newer compilers that diagnose invalid
code in libpdx, one of bufferhub's dependencies.

Bug: 246977679
Change-Id: Iff5f879c2bb5583e789a604d64638e8ce7d2b2c5