Revert "rust: Fix Debug implementation for declare_binder_enum!"
This is an Android Culprit Assistant revert, created as part of a culprit search session (go/aca-get/ea948806-9e80-435d-af82-8bacb80822b9). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Id0fde19ace7d06da3a201e8e63da772a83a1eb89