Revert "rust: Fix Debug implementation for declare_binder_enum!"
Revert "rust: Fix Debug implementation for declare_binder_enum!"

This is an Android Culprit Assistant revert, created as part of a culprit search session (go/aca-get/715a2c84-e5bc-4f67-9e40-7deccce90810). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I2850f3a9214e1b0b4dedacd73115c335a9ce0ba0