Revert "Update IsZygote to return true for webview_zygote"
This is an Android Culprit Assistant revert, created as part of a culprit search session (go/aca-get/85c88b22-1e8a-4687-a5b6-f2c080ff7f11). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I7a48501d957cb2f9b12b9a0511ae8a93bea2724b