Add ability to detect if a DisplayID is valid.
Problem: Currently screencap relies on DisplayId::fromValue to detect
if an id is valid or not. Users can pass any argument, including
zero, which is invalide and result in the error message not being
displayed.

Solution: Tag all types of DisplayID, including Physical displays. Do
not rely on abscence of tag to infer a DisplayId type.

Bug: 254887048
Test: DisplayId_test.cpp
Change-Id: Ief7a6181d037e99054cfb783b5431d7dfa35d2fa