Revert "Do not log an error about ro.sf.lcd_density on emulators"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/2ad9b633-0f08-4955-b6c0-e7fcd2f6a38d). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Iaee1764cdafca50f5606b4c34ef5f0200e51fdc1