Revert "Do not log an error about ro.sf.lcd_density on emulators"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/6a55c392-35c3-42c7-b094-ebf36dce72d0). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: If39d9fb285060a1bfed4ed518ec90488e6df74f1