[rpc_binder] Adjust return when raw socket is not set up
This cl adjusts the return in RpcServer::setupRawSocketServer()
when the socket_fd is not set up. Prior to this cl, errno is
returned. This can be confusing as errno is not always set up in
this case. The current CL fixed this and logged more clear
information in this case.

Test: atest binderRpcTest
Bug: 222479468
Change-Id: Ic9718cbf3cf08befe4044e7f5f021b915b4c4c41