Revert "libbinder_ndk: toString never goes to empty strings."
Revert "update toString checks"

This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/c1e04d15-d7a3-436c-9d7a-61c72906c129). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ia914a304cbd79ce189f4fd55789de759048270b9