Revert "libbinder_ndk: toString never goes to empty strings."
Revert "update toString checks"

This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/2aefd60e-3b33-4fe2-a17a-2d4558f8c2c4). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I15d09e42ffbd228b0bff35043cf8f604fa532d9e