Revert "libbinder_ndk: toString never goes to empty strings."
Revert "update toString checks"

This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/298eb9bb-36ac-4141-81de-a6aedb4d3a05). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I0543359f1272a754fe0105adf3ef3494c2885ebb