Revert "libbinder_ndk: toString never goes to empty strings."
Revert "update toString checks"

This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/2b572586-33e7-4bc8-872a-7c7e12f849d3). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I43b58e6647409368f144e76bab3d610fb2d62724