libbinder: Modify RpcServerTrusty for kernel
This commit modifies RpcServerTrusty to work as a thin
wrapper around RpcServer by adding a port name when building
for the trusty kernel. Unlike userspace, IPC-specific
methods and members are not required since we will only
support the case where the Bn and Bp implementations are
both in the kernel to avoid parsing IPC messages.

Bug: 230134581
Test: build Trusty
Change-Id: I7fa42d132336a497364cbcec1dc7fdd359ead8c0