Remove set_cpuset_policy due to new taskprofiles for SF threads
Now that we have this CL:
SurfaceFlinger: add custom taskprofiles for SF threads

In some platforms, setting a different cpuset (or remove it) and/or
uclamp setting is beneficial to both performance and power. Adding 2
separate profiles for vendor to override surfaceflinger scheduling
policies.

Maybe it's time we got rid of the old interface #set_cpuset_policy
Because the old interface didn't work and was confusing.

Change-Id: Icaedeb17ddf2b14a5113ea84da0bb4e17d4abd74
Signed-off-by: lihaizhou <lihaizhou@coolpad.com>