rpc_binder: Work around RpcServer shutdown deadlock
RpcServer::~RpcServer invokes shutdown() to trigger exit from all
join and session threads. The function waits for the number of
connections to drop down to zero, but this depends on RpcSession
promoting a wp<RpcServer> to sp<RpcServer>. Since this is happening
during the destructor, when the refcount is zero, this pointer
promotion fails. As a result, the list of connections may not be fully
cleared and the thread calling shutdown() will deadlock.

Work around this problem by always calling shutdown() in
ARpcServer_free before decrementing its refcount. This is not ideal:
  * ARpcServer_free should not need to assume that the refcount is being
    dropped to zero,
  * this does not fix the problem for direct users of the C++ class,
but works for all of the existing users.

Bug: 263168076
Test: cleanly shutdown RpcServer with many connections
Change-Id: Ia67a4a839419aafb1bd47fb93ed2e76d56b107c2