Revert "Set CLOEXEC on socket inherited from init"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/1debeebe-6d8e-4b54-8300-9b66146fd3fe). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ia671bf5d13983e376fba10731542f55a1206e8a7