libbinder: Duplicate channel handle in RpcServerTrusty
The tipc library calls close() on the channel handle
when the connection is shut down. RpcServerTrusty wraps
the channel handle in a unique_fd for its own use.
This second copy of the handle needs to be a duplicate
from dup() to prevent a double call to close() on the
same handle.

Bug: 259517277
Test: trusty_stats_test
Change-Id: Ibbab75873ccf43e6d8315202794534e9f89a65c3