Revert "servicemanager: fix lazy service issues"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/6b70434d-4a85-4c2a-8e9e-8d2b3d0f9db7). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ibcbc5544e667fa5fba6c563140f9efcc41ae92fa