Revert "servicemanager: fix lazy service issues"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/167fb045-a91a-4d86-a8bc-5d3258535ade). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ib1530d91eed01c206bf8752b325780c9983b3da2