Revert "servicemanager: fix lazy service issues"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/c31a7194-f2e8-43ba-a931-978f99ba5d93). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ia18248019b06233b02e90b9ee63c6af9884fcd1b