Revert "libbinder: Disable binderRpcTest.vendor and variants"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/7b858501-65d6-4388-afc4-49449a3ca42c). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I078a0b9e787690fbb7b5f421217cd86f9ff4d380