Try erofs when mounting OTA partitions.
We don't know which filesystem type the partitions may have, but
currently only ext4 and erofs are in use.

A more generic approach would be to use libext2_blkid or the blkid
binary to find the filesystem type from the block device itself, since
update_engine doesn't divulge this information. However this is good
enough for now.

Test: m dist
      system/update_engine/scripts/update_device.py out/dist/cf_x86_64_phone-ota-eng.*.zip
  on a Cuttlefish device, and check logcat that ext4 mounts fails but
  erofs ones work.
Bug: 270940551
Bug: 202022157
Change-Id: I77b9472418d7996695c2b28b20d43711d5c4d771