RPC Binder: setMaxOutgoing{Threads,Connections}
This API is updated to emphasize that it is placing a limit
on the number of connections that will be started corresponding
to threads in a remote pool. These connections won't actually
correspond to threads in the process that this API is called.

Documentation has also been updated to clarify a few questions
from an internal discussion forum.

Fixes: 270374393
Test: binderRpcTest
Change-Id: Ia0d9f0d0f42f58a2c63bf506476b33985f091a34