Revert "MotionEvent: Round coordinates to a precision of 0.001"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/7e0f2c8c-3fd8-4252-a17d-1891cc31c206). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ibc0305be18bb02e999c3247a0635d34b98ae49f1