Revert "MotionEvent: Round coordinates to a precision of 0.001"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/78cc5570-3ae9-4a3c-b0ff-be9bafef0519). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Iaf876171b6d6539f652c1c91279ca822a0d05528