Revert "RPC Binder: dropping all binders drops session"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/21af49b0-441f-484d-8548-dd5cda8cd82e). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ide936ffb6ab26d04c2f0c07834ccd7464a1528f0