Update IsZygote to return true for chrome_zygote
This change adds "com.android.chrome_zygote" to the list of processes
that the IsZygote() function will return true for. This is a very
similar change to aosp/2254847, and is also intended to reduce the time
spent in dumpstate's "DUMP TRACES".

While testing another change for backtrace dumps, I noticed that the
only process that still timed out was com.android.chrome_zygote, which
sounded like it should also be ignored.

Test: Check with health/bugreport/capturebugreport that the chrome
      zygote process no longer appears as a timed-out process in
      libdebuggerd_client.
Bug: 276451949
Change-Id: Id2973188d0f8dcf543bc9268bde4a927a563473d