Revert "Include module dependencies under `data`"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/b4533ccf-c85b-4d4b-b511-dd8b5aeda5e7). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ic67b29355bc820aeb1622c55b903fc8b1d94dab7