Revert "Include module dependencies under `data`"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/fb23de0a-5ddc-42ff-adf2-e9ce7db5ff80). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ibb5519dd11deb7da5a4771d9e4704ac48132901d