Revert "libbinder RPC: Trusty always use stable wire proto"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/271c8159-1476-4004-834a-2d7b53ddda81). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I33fb29bc3fd94330302c9812fe95d22d375a3c62