binder: implement Parcel::objectsCount for RPC parcels
From a survey of android code, `objectsCount` is only used to determine
whether a parcel is OK to marshal (i.e. whether is just plain-old-data).
It is correct to consider objects in RPC parcels as also unfit to
marshal.

In `RpcState::transactAddress`, the `data.objectsCount() != 0` check is
no longer relevant. It is now impossible to have kernel binder objects
when `data.isForRpc()` is true.

Bug: 234614894
Test: atest binderRpcTest
Change-Id: Icfe5b74b919ee6b204d7d0ed0b0226178800ff6d