Revert "Do not allow wildcard matching in GL loader."
This reverts commit c034b3a9bc122e4df4f5ebadf45c585612ef80d3.

Original commit message:
```
Previously when ro.hardware.egl and ro.board.platform were not set, the
loader would attempt to load the exact GLES drivers, and if it failed,
it would attempt to use wildcard matching to find the GLES drivers.

However, ro.hardware.egl must be set to point to the GLES drivers and
hence wildcard matching should be disallowed. This patch makes sure the
GL loader no longer uses that path if a device is launched in Android 14
and beyond.
```

Reason for revert: startup regression in b/283393625

Bug: b/283393625
Change-Id: I7bf18d62ea7a15d6d203328fe2c6a453051a7206
Test: hot_startup_com.google.android.calculator-mean