Revert "libbinder_ndk_unit_test: log service test cbs"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/957e5d16-3d7c-422c-9afa-656d7c22c243). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I5630dd6130387846bcc5c57e3b386ff4f7a08485