Remove write permission from file mode of top-level user dirs
https://r.android.com/2620458 is removing the write mode bit from the
top-level user directories on internal storage, in order to make the DAC
consistent with the SELinux policy.

This commit makes the corresponding change to
InstalldNativeService::tryMountDataMirror().  This method creates
/data_mirror/data_ce/$volume_uuid and bind-mounts /data/user onto it,
and similarly for several other directories.  It should not really need
to be changed, but the issue is that for adoptable storage it runs
multiple times, so the modes it uses to prepare the directories with
must match the modes of the corresponding bind-mounted directories.

Bug: 285239971
Test: sm set-virtual-disk true
      sm partition disk:7,392 private
      Verified no error message from tryMountDataMirror() in log
Change-Id: I17dfbe10909b34c2046a4d5b4ffd7764d5ae083b