frameworks/native: Fix multi-display multi-mouse issue
The inputflinger can not handle multi USB mouse for multi display
use case.
An attempt was made to append the input-port-associations.xml
configuration file to vendor/etc/ with official documentation,
but can not handle multi USB mouse for multi display use case.

The reason is that USB mouse can be multiple, and the associated
diaplay ID can be multiple,but PointerController is singleton unique
and only one viewport and one display id are saved in PointerController
So in the case of multiple USB mouse, although the associated display
id is saved in each USB mouse, the target display id and the target
viewport are judged according to the viewport and display id saved in
PointerController,so only one USB mouse can be mapped to one display
other USB mouse can't be mapped to display.

So need to remove the code that determines the display id based on the
controller,and set an associated display id for each CursorInputMapper.
Correct the viewport saved in PointerController based on the associated
viewport before handling mouse events.

Bug: None
Test: atest inputflinger_tests
Change-Id: I907d0e7675df02c1edb42060a3e8fe535feac137
Merged-In: I907d0e7675df02c1edb42060a3e8fe535feac137