Cancel drag and drop when synthesize cancel for all connection
When the policy consume the motion event, the addDragEventLocked will
not invoke, so, the drag flow will keep to next motion stage.
Also, if we start the drag and drop with the second finger, at this time, the policy consume this motion event, because L2543, it will cause system crash at next motion stage.
So, we should cancel drag and drop we the cancel synthesize for all
connection.

Bug: 291181957

Change-Id: Ic5e328f0e445e3271cb1797725c4c00589a6b5e1
Signed-off-by: Linnan Li <lilinnan@xiaomi.corp-partner.google.com>