Revert "[Thread] define the ThreadNetwork hardware feature"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/63b91366-40a8-4770-9127-046c04bbf4ce). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Iba6d075325bcaf1922983b74c2d18ae6ae614c98