Revert "[Thread] define the ThreadNetwork hardware feature"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/64b77cdc-96d9-4ee6-938e-06b77b7cd7dc). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Idb74a9515484e2b6a5039239b47176bd95b81164