Revert "[Thread] define the ThreadNetwork hardware feature"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/fa8d757d-a89d-4975-aa67-0dd35c8977e9). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I9dfd9a4ec725e6c8a975c796a40e4a137cd7b42c