Revert "[Thread] define the ThreadNetwork hardware feature"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/91a986bc-4bf3-4f95-8766-fe604772ce8e). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ic99ee60a11041530e58bca3dde03e40fc96634d6