Revert "[Thread] define the ThreadNetwork hardware feature"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/49b05de9-2f90-4f97-a119-8014bd21e19d). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I6a2efd970603fcd2f01bd10d5fb59b7031e135fe