Use a single otapreopt_chroot invocation for all otapreopt runs.
Instead of taking the dexopt command on the command line it reads a
sequence of commands from stdin.

This changes the interface between otapreopt_script and
otapreopt_chroot, but since the script is in the OTA package while the
otapreopt_chroot command is in the old (currently active) system image,
the script needs to work with the old otapreopt_chroot command too.

The updated script uses `otapreopt_chroot --version` to detect a legacy
binary, which will fail with a "Not enough arguments" error (and there
will also be an sepolicy denial for the use of pipes).

A legacy script with an updated chroot binary won't work either, but
that should only happen in an OTA downgrade when a user exits a beta
program, and in that case the data partition is wiped, so this is moot
anyway.

In a local test on Cuttlefish this reduces the OTA dexopt time with
about 1.5 seconds per package.

Test: DexOptOtaTests, where the system image and the OTA package are
      built with the change
Test: DexOptOtaTests, where only the OTA package is built with the
      change
Bug: 293639539
Change-Id: Ie4e29beaee398428680a0c003f7cb01f2cdd76f9