installd: add enableFsverity
Why installd:

In order to allow the apps to use fs-verity, as the same time avoid the
concern of exposing a risky attack surface in kernel (PKCS#7 signature
verification) to the apps, we need to proxy the FS_IOC_ENABLE_VERITY
ioctl to a trusted process to construct the ioctl argument with minimal
static configuration.

Also, since the ioctl touches an app's file, it requires root privilege
to actually enable fs-verity to the file.

Security considerations:

The service must only enable fs-verity to the given file, if owned by
the caller. The calling uid is provided from the API by trusted
system_server, together with the fd from the app.

Bug: 285185747
Test: Call the API from a local client
Change-Id: I0782a9b23f3817898df0703cfdd9d8670f8fcdfb