Clear unexpected buffer handle of sideband layer
If the buffer handle is not cleared after mediacodec disconnected
and the surface is re-used for another sideband stream. Surfaceflinger
would trigger incorrrect composition type change from sideband to
device when SurfaceView layout changed.

Clear the buffer if the layer was about to set a new sideband
handle.

Bug:291202822

Test: Adjust the screen layout after played 2 different sideband
stream. No composition type change.

Change-Id: Ibd6973134bd48d7d74ef2bf2985ca7f18088f3f1