Revert "libbinder_ndk_unit_test: avoid race"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/9ae36c95-a8f7-4ed0-8200-1fad54a1bf1a). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I0a5c253d57cb9ed978b6a35b92fade69813eda32