Revert "binderRpcBenchmark: test many proxy behavior"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/b6bd4df8-5acd-4d05-afe9-6921858609d1). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I98d6f8566e2d1e1b62029f068e0fbb703495b44a