Revert "Migrate from android::String path functions to std::file..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/faf0fe22-3e66-4858-9d79-8b7dce2acea5). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I04036663691a6aec48fc0ecb1bcde4dc592f1fe4