Revert "binderRpcBenchmark: test many proxy behavior"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/e40affb6-d793-4eab-a136-4cc1159a331e). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I3f85012048c3285f43a125bded093d7a1a227ccb