Revert "binderRpcBenchmark: test many proxy behavior"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/08b99350-555a-4f7d-80d4-0c4592781471). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I8e75758bdcfa3472cf589dcdf591d34237593c81