binder: pass interface descriptor with binder
When sending any binder around, this also sends
the type information. The main benefit of this
is that, you can always check the type of a binder,
even if you aren't able to make a binder call, or
if the binder is dead. Specifically, when binder
proxies leak in Java, this would make sure we're
still able to print the type of the binder.

This adds memory use and reduces transactions, but
it also opens up a variety of other possibilities
or future work:
- we could switch RPC binder to using this same
  optimization and remove uses of locks from
  binder proxies
- we could make interface_cast or Java Stub
  constructors check the type, so we would
  get type errors early
- we could switch writeInterfaceToken to avoid
  writing interface tokens (interface tokens would
  only be written the first time a binder is sent)
- we could move atrace logic from AIDL to libbinder
  so a single atrace call could catch all
  transactions (though, this would also need a way
  to convert code into method name, it would
  remove a huge amount of auto-generated code)
- we could cache descriptors, so that we didn't
  need to store copies of them, when multiple
  binder objects have the same type

Bug: N/A
Test: boot device, binderAllocationLimits
Change-Id: I4849d70b2f5632a89ff7f5d2a8afdd13730e50df