Revert "Rust API to fuzz multiple binders"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/d7a62602-c806-417b-a387-8556fcf6a083). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I25c2eebb83b5862d17d3a59dda7eb62b90b38258