Revert "Add prebuilt_etc for consumerir feature xml"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/6b147428-901b-431f-9883-f2b5aeeb2ea9). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I3e7f2c5abfbb0477dd81419d253d962b5b48c7c4