Added local_downcast function for binder native
Added a local_downcast function to be able to cast
a Binder Native object back to its base object.
WIth this change the cast of a Binder native object
will look like this:

let my_foo = bn_foo.local_downcast::<MyFoo>().unwrap();

This change depends on changes to the AIDL compiler
to use the downcast_rs crate on the autogenerated
Rust code.

Test: CI
Bug: 278780666
Change-Id: I00958df53bb9f052347992ada121d7febfc87dc5