Send cancel events to the correct display when mirroring
When a display is mirrored, such as when screen recording is enabled,
there will be clones of the input windows on the mirrored display that
are on the mirror display.

When there is a stream of input going to the channel through one of
those windows and that stream should be canceled, the generated cancel
event should be sent to the correct window on the correct display.

Bug: 299074463
Test: atest inputflinger_tests
(cherry picked from https://googleplex-android-review.googlesource.com/q/commit:164633818b22bf3485b519e000d48bb33b3b1764)
Merged-In: I7b9eab69c9e9086750cf9fe8a10998a12c30d084
Change-Id: I7b9eab69c9e9086750cf9fe8a10998a12c30d084

NOTE FOR REVIEWERS - errors occurred while applying the patch.
PLEASE REVIEW CAREFULLY.
Errors:
Error applying patch in services/inputflinger/dispatcher/InputDispatcher.cpp, hunk HunkHeader[3958,28->3958,24]: Hunk cannot be applied
Error applying patch in services/inputflinger/tests/InputDispatcher_test.cpp, hunk HunkHeader[4699,6->4699,36]: Hunk cannot be applied

Original patch:
 From 164633818b22bf3485b519e000d48bb33b3b1764 Mon Sep 17 00:00:00 2001
From: Prabir Pradhan <prabirmsp@google.com>
Date: Thu, 12 Oct 2023 23:03:19 +0000
Subject: [PATCH] Send cancel events to the correct display when mirroring

When a display is mirrored, such as when screen recording is enabled,
there will be clones of the input windows on the mirrored display that
are on the mirror display.

When there is a stream of input going to the channel through one of
those windows and that stream should be canceled, the generated cancel
event should be sent to the correct window on the correct display.

Bug: 299074463
Test: atest inputflinger_tests
Change-Id: I7b9eab69c9e9086750cf9fe8a10998a12c30d084
---